[BEAM-1347] Break apart ProcessBundleHandler to use service locator pattern based upon URNs.
This cleans up ProcessBundleHandler and allows for separate improvements of the various PTransform handler factories.
